// src/types/gamedata.type.ts
interface GameData {
  game_id: number;
  start_utc: string;
  duration: number;
  queue: string;
  platform_id: string;
  map_id: number;
  game_mode: string;
  game_version: string;
  participant_id: number;
  champion_id: number;
  champion_name: string;
  position: string;
  win: string;
  kills: number;
  deaths: number;
  assists: number;
  kda_ratio: number;
  kill_participation: number;
  gold_earned: number;
  gold_spent: number;
  gold_per_min: number;
  damage_dealt: number;
  damage_per_min: number;
  damage_to_champ: number;
  damage_champ_per_min: number;
  damage_taken: number;
  vision_score: number;
  item0: string;
  item1: string;
  item2: string;
  item3: string;
  item4: string;
  item5: string;
  item6: string;
  solo_tier: string;
  solo_rank: string;
  solo_lp: number;
  solo_wins: number;
  solo_losses: number;
  flex_tier: string;
  flex_rank: string;
  flex_lp: number;
  flex_wins: number;
  flex_losses: number;
  mastery_level: number;
  mastery_points: number;
  mastery_lastPlayTime: number;
  mastery_pointsSinceLastLevel: number;
  mastery_pointsUntilNextLevel: number;
  mastery_tokens: number;
  final_abilityHaste: number;
  final_abilityPower: number;
  final_armor: number;
  final_attackDamage: number;
  final_attackSpeed: number;
  final_movementSpeed: number;
  final_health: number;
  final_healthMax: number;
  final_lifesteal: number;
  final_omnivamp: number;
  final_power: number;
  final_powerMax: number;
  final_spellVamp: number;
  team_baronKills: number;
  team_dragonKills: number;
  team_towerKills: number;
  team_champKills: number;
  team_riftHeraldKills: number;
  team_inhibitorKills: number;
}

export type { GameData };